поправил ошибки, вынес строковые константы в selenium steam api, вынес работу со стейтом в бд в отдельный класс

+++ Backend/SteamTradeBot.Backend/Models/Abstractions/IConfigurationManager.cs
@@ -4,6 +4,9 @@ namespace SteamTradeBot.Backend.Models.Abstractions;
 
 public interface IConfigurationManager
 {
+    /// <summary>
+    /// Секретный ключ клиента.
+    /// </summary>
     public string ApiKey { get; }
 
     /// <summary>
@@ -20,21 +23,65 @@ public interface IConfigurationManager
     /// Количество продаж предмета за один день.
     /// </summary>
     public int SalesPerDay { get; }
+
+    /// <summary>
+    /// SteamID пользователя.
+    /// </summary>
     public string SteamUserId { get; }
+
+    /// <summary>
+    /// Промежуток цены, при выходе из которого заказ на покупку будет отменен.
+    /// </summary>
     public double FitPriceRange { get; }
+
+    /// <summary>
+    /// Количество страниц листинга цен продажу, которые будут просмотренны.
+    /// </summary>
     public int SellListingFindRange { get; }
 
     /// <summary>
     /// Коэффициент продаж - лежит в промежутке 0:1 - место в списке цена-продажи (чем больше коэффициент, тем ниже будет цена и тем дольше будут покупаться предметы).
     /// </summary>
-    public int SalesRatio { get; }
+    public double SalesRatio { get; }
+
+    /// <summary>
+    /// Период анализа в днях, в рамках которого анализируется график предмета.
+    /// </summary>
     public int AnalysisIntervalDays { get; }
+
+    /// <summary>
+    /// Количество экземпляров предмета в заказе на покупку.
+    /// </summary>
     public int OrderQuantity { get; }
+
+    /// <summary>
+    /// Минимальная цена предмета при формировании списка имен предметов, которые необходимо проанализировать.
+    /// </summary>
     public double MinPrice { get; }
+
+    /// <summary>
+    /// Максимальная цена предмета при формировании списка имен предметов, которые необходимо проанализировать.
+    /// </summary>
     public double MaxPrice { get; }
+
+    /// <summary>
+    /// Размер списка имен предметов, которые необходимо проанализировать.
+    /// </summary>
     public int ItemListSize { get; }
+
+    /// <summary>
+    /// Коммиссия Steam, которая учитывается при поиске оптимального заказа на продажу.
+    /// </summary>
     public double SteamCommission { get; }
+
+    /// <summary>
+    /// Необходимая выгода, которая учитывается при поиске оптимального заказа на продажу.
+    /// </summary>
     public double RequiredProfit { get; }
+
+    /// <summary>
+    /// Коэффициент доступного баланса кошелька стим.
+    /// </summary>
     public double AvailableBalance { get; }
     public Task<bool> RefreshConfigurationAsync(string apiKey, UserConfiguration userConfiguration);
 }
